requirements.txt
python-telegram-bot==20.7
pytube==15.0.0
mega.py==1.0.8
python-dotenv==1.0.0
apscheduler==3.10.4
pymongo==4.6.1 